Workaround for mount_control yaw issue
In ArduPIlot master, the yaw component of the mount_control mavlink
command is not working.  This prevents automatic and free-look yaw from
working properly in Follow, MPCC, Pano, and Zipline smart shots. As a
workaround, I added a condition_yaw mavlink command to handle the yaw
movement.  The mount_control command still handles the gimbal pitch,
which still works properly.  This is a workaround until the problem in
ArduPIlot is found.